jd03_01 Task C, jd03_02 Task A, B
