+++ src/by/it/grechishnikov/jd03_02/TaskB.java
@@ -1,7 +1,5 @@
 package by.it.grechishnikov.jd03_02;
 
-import by.it.grechishnikov.jd03_02.crud.Users;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
